Blending and animator controller set up for delivery

+++ SparkEngine/AnimatorControllerImporter.h
@@ -19,7 +19,7 @@ private:
 		std::string source;
 		std::string target;
 		uint trigger_num;
-		uint blend;
+		float blend;
 	};
 
 public:
